pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "neurosdk"
version = "0.1.0"
description = "The Python Neuro SDK"
authors = [
  { name="vanorsigma", email="blackhole@blackhole.com" },
]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "websocket-client==1.8.0",
    "dataclasses-json==0.6.7",
    "pygame==2.6.1",
    "rel==0.4.9.19",
    "websockets==14.2",
    "marshmallow-jsonschema==0.13.0"
]
